修改成立年限为小数，且保留小数点后一位

+++ src/main/java/com/idc/modules/entity/BrandBasRevPerf.java
@@ -79,7 +79,7 @@ public class BrandBasRevPerf extends Model<BrandBasRevPerf> {
      * 1.企业成立年限
      */
     @JsonProperty("企业成立年限")
-    private int chenglnx;
+    private BigDecimal chenglnx;
     /**
      * 2.营业收入(上一年)
      */
